[Next] layout崩れたまま

+++ srcs/requirements/wordpress/tools/setup.sh
@@ -10,39 +10,38 @@
 mkdir -p /run/php
 mv /tmp/test.html /var/www/html/
 
-wp cli update --allow-root
-
 sleep 10
 
-wp config create	--path=/var/www/html \
-					--allow-root \
-					--dbname=$WORDPRESS_DB_NAME \
-					--dbuser=$WORDPRESS_DB_USER \
-					--dbpass=$WORDPRESS_DB_PASSWORD \
-					--dbhost=$WORDPRESS_DB_HOST
+if [ ! -f /var/www/html/wp-config.php ]; then
 
-wp core install		--allow-root \
-					--url=$WORDPRESS_URL \
-					--title=$WORDPRESS_TITLE \
-					--admin_user=$WORDPRESS_ADMIN_USER \
-					--admin_password=$WORDPRESS_ADMIN_PASSWORD \
-					--admin_email=$WORDPRESS_ADMIN_EMAIL \
-					--skip-email \
-					--path=/var/www/html/
+	wp core download --path=/var/www/html --locale=ja --allow-root
+	wp config create	--path=/var/www/html \
+						--allow-root \
+						--dbname=$WORDPRESS_DB_NAME \
+						--dbuser=$WORDPRESS_DB_USER \
+						--dbpass=$WORDPRESS_DB_PASSWORD \
+						--dbhost=$WORDPRESS_DB_HOST
 
-wp user create		--allow-root \
-					$WORDPRESS_USER1 \
-					$WORDPRESS_USER1_EMAIL \
-					--user_pass=$WORDPRESS_USER1_PASSWORD \
-					--role=author \
-					--path=/var/www/html/
+	wp core install		--allow-root \
+						--url=$WORDPRESS_URL \
+						--title=$WORDPRESS_TITLE \
+						--admin_user=$WORDPRESS_ADMIN_USER \
+						--admin_password=$WORDPRESS_ADMIN_PASSWORD \
+						--admin_email=$WORDPRESS_ADMIN_EMAIL \
+						--skip-email \
+						--path=/var/www/html/
 
-# wp option update	--allow-root \
-# 					--path=/var/www/html/ \
-# 					home $WORDPRESS_URL
-
-# wp option update	--allow-root \
-# 					--path=/var/www/html/ \
-# 					siteurl $WORDPRESS_URL
+	wp user create		--allow-root \
+						$WORDPRESS_USER1 \
+						$WORDPRESS_USER1_EMAIL \
+						--user_pass=$WORDPRESS_USER1_PASSWORD \
+						--role=author \
+						--path=/var/www/html/
+	# wp plugin install	--allow-root \
+	# 					--activate \
+	# 					--path=/var/www/html/ \
+	# 					akismet \
+	# 					hello
+fi
 
 exec "$@"
